[*] Fixes #373. Tightened up TokenizedString public APIs. THIS IS A BREAKING CHANGE
Parse() and ToString() used to do the same thing, but now they are different.

Parse() no longer returns anything, and if any errors occur, will be cached internally.

ToString() will display the parsed string, or throw an exception if it is not parsed, or include the message if it failed to parse. In short, TokenizedStrings only throw exceptions now when attempted to be used, not when they are parsed, unless a second parse attempt is made.

TokenizedString comparisons now use the hash calculated at construction, so no parsing is required.

Ideally, TokenizedStings should be created during Init() or patching stages. These will be automatically parsed in the parsing stage, and are available for use in the build stage, with ToString().

TokenizedString.ParseAndQuoteIfNecessary() has been replaced with ToStringQuoteIfNecessary().

C.PreprocessorDefinitions now internally store TokenizedStrings, not strings.

Test harness has been changed so that Native build mode runs repeat their build command, but without a clean, as incremental builds run down different codepaths internally. This is to give wider test coverage.

Added ProceduralHeaderTest2, which exercises the C.ProceduralHeaderFileFromToolOutput module type, again increasing test coverage.

Added PublishingSymlinkTest1, which exercises publishing a symbolic link.

TokenizedString allows registering custom post-functions (with a unique name). Publisher uses this to add 'readlink' that reads the symbolic link's target at parse time.

Added DoNotCompileTest1, exercising the PerformCompilation boolean property on C.ObjectFileBase.

Added VersionScriptTest1, to exercise Gcc Ld version scripts.

Now that Native build tests are repeated without a clean, the Installer generation for DMGs on macOS fail, because 'hdiutil convert' expects the output DMG not to exist. Fixed this by writing to a temporary location, and then moving that to the expected location

It was noticed during profiling that the new TokenizedString code was significantly slower than the original, due to calls to get the stack trace. In the new code, this can be called multiple times (creation and parsing). This was done to aid debuggability in error messages. It is now disabled by default, but can be re-enabled at the cost of a slower build (particularly on macOS) by passing --withstacktrace to bam