First attempt at retargeting to DotNetCore v2.1
Retargets to VisualStudio 2017. New .csproj format.

Runtime compilation changed to target the Roslyn compiler API, since the old CSharp compilation API is not supported on dotnetcore.

Running bam now requires a prefix of 'dotnet'. Updated env.bat and env.sh a bit, but not happy with the results there.
Test harness doesn't run because of this.

Only tested on Windows so far.