[packages,C] Fixes #499. Added support for delegating warning suppression.
Added C.SuppressWarningsDelegate abstract class, and C.CCompilableModuleContainer.SuppressWarningsDelegate function, as an aid to delegating responsibility for warning suppressions away from the main build script. Build scripts are then more succinct for the details on the build, rather than bogging down with details of making it build cleanly. The intended use for this is for external code which cannot be fixed.