[Core] Fixes #526
Downloading and extracting package source archives now requires that a Module from the Package has been used in the build. This means that there may be packages with source archives that don't get downloaded on some platforms.
Note that the concurrent download seen in previous alphas of v2.0 is no longer present, since the downloads must now be deferred until a sequential part of the build step.