[*] Fixes #587. No longer need to pass a parent Module to Module.Init()
BREAKING CHANGE.
This change affects every Module class. The use of the parent Module is the exception rather than the rule, and there are alternative methods to get the information if it is needed.
Added Bam.Core.Graph.ModuleStack property in order to peek at the hierarchy of parent Modules. This can be done in Init() calls.
Removed Bam.Core.Module.GetEncapsulatingReferencedModule. Access to the encapsulating Module is now through the readonly Module.EncapsulatingModule property.
The encapsulating Module is now calculated after the Module preInitCallback but before Init().
The following Module macros are no longer available:
* parentmodulename
* encapsulatedparentmodulename
* encapsulatingmodulename
* encapsulatingbuilddir
They can be obtained via the actual Modules accessed by the new properties.